This appears to be working with the video files.
It copies all the tags from the original mp4 to get all the camera data.
However, the local time is saved without the offset and causes issues
uploading to Immich.  This update then uses the UTC equivalent from the
original metadata and pastes that in the place Immich looks for the
timestamp.  Since it's UTC it aligns with the QuickTime standard and
shows in the timeline correctly.  Whew... Now for GPS.. :-/